chapter_003/src/main/java/ru/job4j/phonedictionary/PhoneDictionary.java
package ru.job4j.phonedictionary;
/**
 * @author Dmitry Kolganov (dmk78@inbox.ru)
 * @version $Id$
 * @since 0.1
 */

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

public class PhoneDictionary {
    private List<Person> persons = new ArrayList<Person>();

    public void add(Person person) {
        this.persons.add(person);
    }

    /**
     * Вернуть список всех пользователей, который содержат key в любых полях.
     *
     * @param key Ключ поиска.
     * @return Список подощедщих пользователей.
     */
    public List<Person> find(String key) {
        return persons.stream().filter(person -> person.getName().equals(key) | person.getSurname().equals(key)
                | person.getAddress().equals(key) | person.getPhone().equals(key)).collect(Collectors.toList());
    }
}